backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
redis==5.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.4